fix: add link to add new retro board text on dashboard
