+++ frontend/src/components/Dashboard/RecentRetros/partials/EmptyBoards/styles.tsx
@@ -19,19 +19,6 @@ const EmptyBoardsText = styled(Text, {
 	mb: '48px'
 });
 
-const EmptyBoards: React.FC = () => {
-	return (
-		<StyledBox elevation="1" justify="center" align="center" direction="column">
-			<StyledImage />
-			<EmptyBoardsText size="md" css={{ mt: '$24', textAlign: 'center' }}>
-				You have not participated in any retro yet.
-				<br />
-				<Text weight="medium" underline>
-					Add a new retro board
-				</Text>{' '}
-				now.
-			</EmptyBoardsText>
-		</StyledBox>
-	);
-};
-export default EmptyBoards;
+const StyledNewBoardLink = styled('a', Text, { cursor: 'pointer' });
+
+export { EmptyBoardsText, StyledBox, StyledImage, StyledNewBoardLink };
